requirements.txt

azure-functions
